[RSDK-9960] Have a more obvious error when you can't detect faces in the known images (#23)
One of these files was the problem, and the other likely was not, but I fixed them both (I think they were originally copies of each other). 

I recommend going through this commit-by-commit: I've also refactored some other stuff to simplify the control flow and flatten the code. I made the same refactorings in both files to preserve their similarity (in the hopes of one day refactoring the shared parts so that we only need to write bugfixes once).

My hope is that with this PR, Julie's inscrutable `AttributeError` becomes a `NoFacesDetectedError`, which is easier to understand/fix.

I had to rebase a copy of this branch on top of Robin's `add_torchvision_build` branch, then `mkdir src/models`, then copy in a whole bunch of models to that directory that I found on one of the demo machines. I also needed to copy a database and a directory tree with images in it off of the demo machine. After that, `rm build/pyinstaller_dist/main && make pyinstaller` takes roughly 12 minutes to run no matter how small a change I've made (and without the `rm` it doesn't notice that anything needs to be remade at all). but with all of that, it builds and seems to work!

- When all pictures of people have people in them, things start up normally.
- When some pictures don't have people but all directories have at least 1 valid picture, we log which images we can't find things in, and then start up normally.
- When some directory doesn't have any valid pictures at all, we get into a reconfigure/crash loop with an error describing which directory doesn't have any pictures in it. 
- Same behavior for the directories of faces: we log any images where we can't find a face, and get into a reconfigure loop if there's a directory with no faces anywhere.


* raise an exception if we never found any faces for a known embedding

* tweak: use generator comprehension to simplify if statement

* tweak: flip if statement in anticipation of simplifying control flow

* tweak: continue in if statement, then unindent else statement

* same tweak to the copy of the if statement

* same inversion of if statement

* same use of continue to flatten code

* raise the same exception if no faces are ever found

* fix incorrect type annotation

* check for None in optional value

* split NoFacesDetectedError and NoPersonDetectedError separately

* fix docstring copypasta

* tweak error messages

* add more logging